+++ src/vue/fields/InputField.vue
@@ -78,6 +78,7 @@ export default {
     errorMessage: { type: String, default: undefined },
     whiteAutofill: { type: Boolean, default: true },
     type: { type: String, default: undefined },
+    trim: { type: Boolean, default: true },
   },
 
   data: () => ({
@@ -123,6 +124,10 @@ export default {
         document.removeEventListener('keyup', this.detectCapsLock)
 
         if (!this.value) this.isCapsLockOn = false
+      }
+
+      if (this.trim) {
+        this.$emit(EVENTS.input, event.target.value.trim())
       }
     },